requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.7
django-filter==23.3
djangorestframework==3.14.0
importlib-metadata==6.8.0
Markdown==3.5.1
numpy==1.24.4
opencv-python==4.8.1.78
packaging==24.0
Pillow==10.1.0
psycopg2==2.9.9
pytesseract==0.3.10
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
zipp==3.17.0
